requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
pymysql
mysql-connector-python
python-dotenv
dill
flask
#-e .